backend/pyproject.toml
[project]
name = "career-radar-backend"
version = "0.1.0"
description = "Career Radar MVP backend"
requires-python = ">=3.12"
dependencies = [
  "fastapi>=0.111",
  "uvicorn[standard]>=0.30",
  "sqlalchemy>=2.0",
  "psycopg[binary]>=3.1",
  "alembic>=1.13",
  "pydantic-settings>=2.3",
  "python-multipart>=0.0.9",
  "pandas>=2.2",
  "apscheduler>=3.10",
  "httpx>=0.27",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.2",
  "pytest-asyncio>=0.23",
  "ruff>=0.5",
]

[tool.pytest.ini_options]
pythonpath = ["."]

[tool.ruff]
line-length = 100

[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"
